// src/components/home/__generated__/NFTCollectionsTabScreenQuery.graphql.ts
export type NFTCollectionsTabScreenQuery$variables = {
  ownerAddress: string;
};
export type NFTCollectionsTabScreenQuery$data = {
  readonly portfolios: ReadonlyArray<{
    readonly id: string;
    readonly nftBalances: ReadonlyArray<{
      readonly ownedAsset: {
        readonly collection: {
          readonly name: string | null;
        } | null;
        readonly description: string | null;
        readonly id: string;
        readonly name: string | null;
        readonly nftContract: {
          readonly address: string | null;
        } | null;
        readonly smallImageUrl: string | null;
        readonly tokenId: string | null;
      } | null;
    } | null> | null;
    readonly ownerAddress: string;
  } | null> | null;
};
export type NFTCollectionsTabScreenQuery = {
  response: NFTCollectionsTabScreenQuery$data;
  variables: NFTCollectionsTabScreenQuery$variables;
};